# apps/api/openapi_spec.yaml
openapi: 3.0.3
info:
  title: Nova Universe API
  description: |
    Enterprise-grade ITSM platform API with comprehensive ticket management,
    user management, and AI-powered automation capabilities.
    
    ## Authentication
    This API uses JWT Bearer token authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <token>
    ```
    
    ## Error Handling
    All errors follow RFC 7807 Problem Details for HTTP APIs standard.
    
    ## Rate Limiting
    API calls are rate limited to 1000 requests per hour per user.
    
  version: 2.0.0
  contact:
    name: Nova Universe API Support
    email: api-support@nova-universe.com
    url: https://docs.nova-universe.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.nova-universe.com
    description: Production server
  - url: https://staging.api.nova-universe.com  
    description: Staging server
  - url: http://localhost:3000
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login endpoint
...
paths:
...
  /auth/login:
    post:
      summary: User login
      description: Authenticate a user and return a JWT token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              required:
                - username
                - password
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        '401':
          description: Invalid credentials
...